requirements.txt
coreapi==2.3.3
django-filter==2.1.0
django-simple-captcha==0.5.11
Django==2.2
djangorestframework-recursive==0.1.2
djangorestframework==3.9.3
Markdown==3.1
Pillow==6.0.0
